Improve login/logout btn
