+++ package.json
@@ -14,6 +14,7 @@
     "lint": "eslint . --ext .vue,.js,.jsx,.cjs,.mjs,.ts,.tsx,.cts,.mts --fix --ignore-path .gitignore"
   },
   "dependencies": {
+    "@vueuse/head": "^1.1.15",
     "axios": "^1.3.4",
     "dotenv": "^16.0.3",
     "pinia": "^2.0.28",
