+++ src/assets/stylesheets/partials/base/_themes.scss
@@ -26,3 +26,35 @@ $dark-theme: (
   success-color: #11bdb3,
 );
 
+$default-theme: (
+  // Grid
+  global-grid: 1280px,
+  global-margin: 20px,
+
+  // Font
+  font-size: 20px,
+  line-height: 1.26,
+
+  // Easings
+  ease: ease-in-out,
+  bezier: cubic-bezier(0.175, 0.885, 0.32, 1.275),
+
+  // Spacing
+  spacing: 20px,
+
+  // Radius
+  radius: 30px,
+  radius-md: 20px,
+  radius-sm: 15px,
+
+  // Shadow
+  shadow: rgba(0, 0, 0, 0.05) 0px 0px 0px 1px,
+
+  // breakpoints
+  large: 1400px,
+  desktop: 1280px,
+  medium: 1124px,
+  tablet: 860px,
+  small: 600px,
+  phone: 400px,
+);
